requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
transformers>=4.30.0
torch>=2.0.0
streamlit>=1.25.0
requests>=2.28.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
python-multipart>=0.0.5
pydantic>=1.10.0,<2.0.0
pytest>=7.0.0
python-dotenv>=0.19.0 